requirements.txt
selenium>=4.8.0
webdriver-manager>=3.8.4
pytest>=7.0.0
faker>=18.0.0
appium-python-client>=3.5.0
python-dotenv>=1.0.0
pytest-html>=3.2.0
pytest-xdist>=3.1.0
allure-pytest>=2.9.45